Test class variables
This does not work in cases where the variable is declared out side a
namespace. Currently, we generate those variables in the project package
directly, but this prevents use from including the required (Class) type
defined in a child package of that project package.

An option could be a distinct package hierarchy for types and instances
below the project package.

Ref. #52